bug fixes continue
